Fix backup target was not properly set when specified by command line
When target path was specified by command line, it was never honored. Files were saved under "/save_config" and "/log" directories in root path and did not survive a reboot. No data loss was caused by this bug. If path was properly set in script, backups were honored as expected. This commit fixes the issue + adds a few enhancements:

- support paths with trailing space char
- user name variable to modify user with rar binary installed (optional)
- disable globbing and splitting of positional arguments if some generic wildcards were supplied as argument (exp. '*')
- do not allow use of non allowed chars in arguments "&;({<>`|*"
- cosmetic log formatting output changes when  backup was aborted